Example from Twitter
This doesn’t fix the entire issue. The bottom two sections still have
the issue, but their min-height could be modified (as sort of a hack).
The proper solution would be to fully use the table layout (using CSS),
but I don’t have the time right now to show you the entire process.